0.02.4
Fixed Main menu to work with trigger rather grip. Want to establish early the player uses trigger to activate while grip is used to grip objects or interactable points.

Main menu will show path, when a new game is selected the menu will disappear and allow the player to move to give a quick smooth transition into the game. The player will not need to walk far for the events leading to the tutorial room to happen.